Fix dropdown menu bug for guest selection
